todo funcionando ok

+++ routes/adotante.py
@@ -1,24 +1,24 @@
 from fastapi import APIRouter, Depends, Query, HTTPException
 from sqlmodel import Session, select
 from database import get_session
-from modelos.adotante import Adotante
+from modelos.adotante import Adotante, AdotanteBase
 
 router = APIRouter(prefix="/adotantes", tags=["Adotantes"])
 
 # CREATE
 @router.post("/")
-def create_adotante(adotante: Adotante, session: Session = Depends(get_session)):
+def create_adotante(adotante: AdotanteBase, session: Session = Depends(get_session)):
+    novo = Adotante(**adotante.model_dump())
     print("Creating adotante:", adotante)
-    session.add(adotante)
+    session.add(novo)
     session.commit()
-    session.refresh(adotante)
-    return adotante
+    session.refresh(novo)
+    return novo
 
-# READ - Listar todos adotantes (paginação)
+# READ - Listar todos adotantes
 @router.get("/")
 def read_adotantes(offset: int = 0,limit: int = Query(default=10, le=100),session: Session = Depends(get_session)):
-    stmt = select(Adotante).offset(offset).limit(limit)
-    return session.exec(stmt).all()
+    return session.exec(select(Adotante).offset(offset).limit(limit)).all()
 
 #  c) Buscas por texto parcial - Buscar adotante pelo nome
 @router.get("/buscar/nome")
@@ -41,7 +41,7 @@ def read_adotante(adotante_id: int, session: Session = Depends(get_session)):
 
 # UPDATE
 @router.put("/{adotante_id}")
-def update_adotante(adotante_id: int, adotante: Adotante, session: Session = Depends(get_session)):
+def update_adotante(adotante_id: int, adotante: AdotanteBase, session: Session = Depends(get_session)):
     db_adotante = session.get(Adotante, adotante_id)
     if not db_adotante:
         raise HTTPException(status_code=404, detail="Adotante não encontrado")
@@ -53,14 +53,18 @@ def update_adotante(adotante_id: int, adotante: Adotante, session: Session = Dep
     session.refresh(db_adotante)
     return db_adotante
 
-# DELETE
 @router.delete("/{adotante_id}")
 def delete_adotante(adotante_id: int, session: Session = Depends(get_session)):
     adotante = session.get(Adotante, adotante_id)
     if not adotante:
         raise HTTPException(status_code=404, detail="Adotante não encontrado")
 
+    if adotante.adocoes:
+        raise HTTPException(
+            status_code=400,
+            detail="Adotante possui adoções registradas e não pode ser removido"
+        )
+
     session.delete(adotante)
     session.commit()
     return {"ok": True}
-
